capstone_src/requirements.txt
altair==5.1.2
annotated-types==0.6.0
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
blinker==1.6.3
blis==0.7.11
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.1.4
confection==0.1.3
cymem==2.0.8
debugpy==1.8.0
dnspython==2.4.2
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.0/en_core_web_md-3.7.0-py3-none-any.whl#sha256=d37c3d60738be409a04f83a51421075b94fc02efad99bb8e62ec9018b1f99cb0
executing==2.0.0
GitPython==3.1.40
importlib-metadata==6.8.0
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.4.0
jupyter_core==5.4.0
langcodes==3.3.0
markdown-it-py==3.0.0
matplotlib-inline==0.1.6
mdurl==0.1.2
murmurhash==1.0.10
nest-asyncio==1.5.8
packaging==23.2
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.17.0
plotly-express==0.4.1
preshed==3.0.9
prompt-toolkit==3.0.39
psutil==5.9.6
pure-eval==0.2.2
pyarrow==13.0.0
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
statsmodels==0.14.0
streamlit==1.27.2
thinc==8.2.1
threadpoolctl==3.2.0
toml==0.10.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
typer==0.9.0
typing_extensions==4.8.0
tzlocal==5.1
validators==0.22.0
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.8
weasel==0.3.3
zipp==3.17.0
anyio==3.7.1
decorator==5.1.1
exceptiongroup==1.1.2
gitdb==4.0.10
h11==0.14.0
httpcore==0.17.3
idna==3.4
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
numpy==1.25.1
pandas==2.0.3
protobuf==4.23.4
pydeck==0.8.0
pymongo==4.4.1
Pympler==1.0.1
python-dateutil==2.8.2
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.0
rich==13.4.2
rpds-py==0.9.2
smmap==5.0.0
sniffio==1.3.0
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tzdata==2023.3
urllib3==2.0.4
seaborn
bokeh


